fix: 런게임 충돌 해결 및 v30 유물 아이콘 반영
- v30 유물 카탈로그와 Relic Sheet 300 Addressables 매핑 유지
- DevButtonAttribute 디버그 프리팹과 누락 스크립트 복원
- BattleFlowController의 다시 표식 유물 흐름 보존
- STATUS.md 충돌을 최신 v30 상태 기준으로 정리

+++ Assets/_Project/Scripts/UI/Leaderboard/LeaderboardModels.cs
@@ -23,7 +23,7 @@ namespace SlotRogue.UI.Leaderboard
     public sealed class LeaderboardSymbolCount
     {
         public string Symbol;
-        public int Count;
+        public float Count;
 
         public LeaderboardSymbolCount()
         {
@@ -31,10 +31,10 @@ namespace SlotRogue.UI.Leaderboard
             Count = 0;
         }
 
-        public LeaderboardSymbolCount(string symbol, int count)
+        public LeaderboardSymbolCount(string symbol, float count)
         {
             Symbol = symbol ?? string.Empty;
-            Count = Math.Max(0, count);
+            Count = Math.Max(0f, count);
         }
 
         public bool TryGetSymbol(out SlotSymbolType symbol)
